The set of finished files has been moved from the Job object to the job runner to avoid it being serialised to disk each time the service instance is updated.  It is re-created at regular intervals (currently 2s) by the CheckOutputFiles thread launched by the job runner.  The threads involved in file downloading on the server (downloadOutputFile method in the GetOperationsController class) check the set at regular intervals to see if they can come to an end and delete the file.
Another small change at the client: The status updater thread launched by the job runner no longer maintains a list of files that have already been downloaded.  Before, this list would be checked before deciding whether to launch a downloader thread for a particular file.  This led to performance problems when the list became large, but it is not really necessary to have such a list at all as long as the files are deleted (or have their names changed) when output has finished at the server.

git-svn-id: svn://svn.code.sf.net/p/jstyx/code/trunk@865 aeb1b72f-4910-0410-8ae4-b0e1c25358ee