New non-recursive method for deleting an old working directory with the same service instance ID as a newly generated ID.  This is used instead of the recursive method to prevent the unintentional deletion of subdirectories that are symbolic links to persistent directories.  The non-recursive method does not delete subdirectories at all.  A better solution would be to enable the recursive method to delete symbolic links rather than follow them.
In another change related to newly generated service instance IDs, old working directories of the same name are renamed by appending "-old" plus an integer >=0 to the original name while awaiting deletion.  The integer in the name is incremented if an even older directory with the same ID still exists.

git-svn-id: svn://svn.code.sf.net/p/jstyx/code/trunk@886 aeb1b72f-4910-0410-8ae4-b0e1c25358ee