A file can now be classified as ready for download if output to it has finished.  Previously, a file had to be append-only or be at the end of the run before downloading could commence. The most significant change to the code is the removal of isOutputFinished() from the job runner object.  This method is replaced by the isOutputFinished() method in the OutputFile class.  This change is necessary to allow each OutputFile object to decide for itself whether or not output to the file has finished.  Unlike the job runner version, This method can be used by OutputFile.isReadyForDownload().
The procedure for deciding whether or not output to a file has finished depends on the service type.  Currently a procedure is only defined for jobs of type "local" and "sge", which use the deleteAfter file attribute as the basis for the decision.  The return value for other service types (yet to bw written) is false.

git-svn-id: svn://svn.code.sf.net/p/jstyx/code/trunk@887 aeb1b72f-4910-0410-8ae4-b0e1c25358ee